:sparkles: 리뷰 제거 전 존재 하는 리뷰인지 확인 후 예외 처리

+++ server/wonprice/src/main/java/main/wonprice/domain/member/service/WishService.java
@@ -1,21 +1,24 @@
 package main.wonprice.domain.member.service;
 
+import lombok.AllArgsConstructor;
 import main.wonprice.domain.member.entity.Member;
 import main.wonprice.domain.member.entity.Wish;
 import main.wonprice.domain.member.repository.WishRepository;
+import main.wonprice.exception.BusinessLogicException;
+import main.wonprice.exception.ExceptionCode;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
+@AllArgsConstructor
 public class WishService {
 
     private final WishRepository wishRepository;
+    private final MemberService memberService;
 
-    public WishService(WishRepository wishRepository) {
-        this.wishRepository = wishRepository;
-    }
 
     public Wish addWish(Wish wish) {
         return wishRepository.save(wish);
@@ -27,6 +30,14 @@ public class WishService {
     }
 
     public void removeWish(Long wishId) {
+
+        Optional<Wish> findWish = wishRepository.findById(wishId);
+
+        if (!findWish.isPresent()) {
+            throw new BusinessLogicException(ExceptionCode.WISH_NOT_FOUND);
+        }
+
+        memberService.validateOwner(findWish.get().getMember().getMemberId());
         wishRepository.deleteById(wishId);
     }
 }
